+++ dropwizard-guicey/src/main/java/ru/vyarus/dropwizard/guice/test/client/builder/track/RequestData.java
@@ -13,7 +13,6 @@ import ru.vyarus.dropwizard.guice.debug.util.RenderUtils;
 import ru.vyarus.dropwizard.guice.module.installer.util.StackUtils;
 import ru.vyarus.dropwizard.guice.test.client.ResourceClient;
 import ru.vyarus.dropwizard.guice.test.client.TestClient;
-import ru.vyarus.dropwizard.guice.test.client.TestRestClient;
 import ru.vyarus.dropwizard.guice.test.client.builder.TestClientDefaults;
 import ru.vyarus.dropwizard.guice.test.client.builder.TestClientRequestBuilder;
 import ru.vyarus.dropwizard.guice.test.client.builder.track.impl.BuilderTracker;
@@ -46,7 +45,6 @@ public class RequestData {
             TestClientRequestBuilder.class,
             TestClientDefaults.class,
             TestClient.class,
-            TestRestClient.class,
             ResourceClient.class,
             RequestModifierSource.class,
             TargetTracker.class,